Add support for retrieving solar panel data from Hoymiles.

* Add the Hoymiles service
* Update the documentation
* Add a depend on the `md-5` crate for password hashing

Reviewed-on: https://git.luon.net/paul/solar-grabber/pulls/6

+++ src/services/my_autarco.rs
@@ -119,8 +119,8 @@ impl super::Service for Service {
         };
 
         Ok(Status {
-            current_w: api_power.pv_now,
-            total_kwh: api_energy.pv_to_date,
+            current_w: api_power.pv_now as f32,
+            total_kwh: api_energy.pv_to_date as f32,
             last_updated,
         })
     }
